responsive web site
Both Flexbox and Grid can be relatively easy to use once you understand the basics of how they work.

Flexbox is a simpler layout system, it has a more limited set of properties and is best suited for one-dimensional layout. It's relatively easy to align items within a container using Flexbox.

Grid, on the other hand, is a more complex layout system, it has a more extensive set of properties and is best suited for two-dimensional layout. It's relatively easy to create a grid-based layout using Grid, but it can take some time to master all its properties and understand how they work together.

Both Flexbox and Grid have good browser support and are widely used in web development. There are also many resources available online such as tutorials, guides, and code examples to help you get started with using Flexbox and Grid.